Add multi page functionality
